Capital Item Volume Validation
Changed the code to better handle capital items.  All capital items in
game are exactly 4000 or 8000 in volume, except for a handful of officer
modules (which this validation will not catch).  The validation is
skipped if the ship is a capital (since capital items are allowed) or if
there are no requirements attached to the ship (since we don't know if
it's a capital or not, we can't really validate it either way).

Tests need to be rewritten to handle that, will update the PR with the
updated tests.